Move Menu to components folder
